Update README.md
This is a function for the Birthday Problem! If you have n people together and equally distributed birthdays 365 days a year, what is the probability that at least two people share a birthday? This function randomly generates a number of birthdays and determines how many matches there are. The function contains num_people corresponding to the number of people together (default 23). The function returns a single value of the number of days that have a match. Additionally, a simulation simulated 100,000 iterations of this process with different amounts of num_people. You can see the distribution of the number of matches of birthdays with the provided histograms.